release version 2.1.0

+++ Models/SerpGoogleFinanceExploreLiveAdvancedResultInfo.cs
@@ -1,4 +1,5 @@
 using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 using DataForSeo.Client.Models;
 
@@ -7,11 +8,6 @@ namespace DataForSeo.Client.Models
 
     public class SerpGoogleFinanceExploreLiveAdvancedResultInfo 
     {
-
-        /// <summary>
-        /// keyword received in a POST array
-        /// <br/>the keyword is returned with decoded %## (plus character ‘+’ will be decoded to a space character)
-        /// </summary>
         [JsonProperty("keyword", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public string Keyword { get; set; }
 
@@ -20,79 +16,26 @@ namespace DataForSeo.Client.Models
         /// </summary>
         [JsonProperty("type", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public string Type { get; set; }
-
-        /// <summary>
-        /// search engine domain in a POST array
-        /// </summary>
         [JsonProperty("se_domain", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public string SeDomain { get; set; }
-
-        /// <summary>
-        /// location code in a POST array
-        /// </summary>
         [JsonProperty("location_code", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public int? LocationCode { get; set; }
-
-        /// <summary>
-        /// language code in a POST array
-        /// </summary>
         [JsonProperty("language_code", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public string LanguageCode { get; set; }
-
-        /// <summary>
-        /// direct URL to search engine results
-        /// <br/>you can use it to make sure that we provided accurate results
-        /// </summary>
         [JsonProperty("check_url", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public string CheckUrl { get; set; }
-
-        /// <summary>
-        /// date and time when the result was received
-        /// <br/>in the UTC format: “yyyy-mm-dd hh-mm-ss +00:00”
-        /// <br/>example:
-        /// <br/>2019-11-15 12:57:46 +00:00
-        /// </summary>
         [JsonProperty("datetime", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public string Datetime { get; set; }
-
-        /// <summary>
-        /// autocorrection of the search engine
-        /// <br/>if the search engine provided results for a keyword that was corrected, we will specify the keyword corrected by the search engine and the type of autocorrection
-        /// </summary>
         [JsonProperty("spell", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public SpellInfo Spell { get; set; }
-
-        /// <summary>
-        /// search refinement chips
-        /// <br/>in this case, the value will be null
-        /// </summary>
         [JsonProperty("refinement_chips", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public RefinementChipsInfo RefinementChips { get; set; }
-
-        /// <summary>
-        /// types of search results in SERP
-        /// <br/>contains types of search results (items) found in SERP;
-        /// <br/>possible item types: google_finance_hero_groups, google_finance_interested, google_finance_news, google_finance_earnings_calendar, google_finance_most_followed, google_finance_market_trends, google_finance_people_also_search
-        /// </summary>
         [JsonProperty("item_types", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public IEnumerable<string> ItemTypes { get; set; }
-
-        /// <summary>
-        /// total number of results in SERP
-        /// </summary>
         [JsonProperty("se_results_count", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public long? SeResultsCount { get; set; }
-
-        /// <summary>
-        /// the number of results returned in the items array
-        /// </summary>
+        public int? SeResultsCount { get; set; }
         [JsonProperty("items_count", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
-        public long? ItemsCount { get; set; }
-
-        /// <summary>
-        /// market indexes related to the market trends element
-        /// <br/>possible type of items: google_finance_asset_pair_element, google_finance_market_instrument_element, google_finance_market_index_element
-        /// </summary>
+        public int? ItemsCount { get; set; }
         [JsonProperty("items", Required = Required.Default, NullValueHandling = NullValueHandling.Ignore)]
         public IEnumerable<BaseSerpApiGoogleFinanceElementItem> Items { get; set; }
